Changing page when creating project works + disabling horizontal scrolling.
