+++ Assets/Scripts/Domain/Common/GameObjectRecord.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ebe5a2521a1d4a8fac94506e29d86a1e
+timeCreated: 1682963189
