+++ Assets/Scripts/UI/DevPage/DevPages.cs
@@ -1,0 +1,9 @@
+namespace UI.DevPage
+{
+    public enum DevPages
+    {
+        Main,
+        NewProject,
+        RunningProjects
+    }
+}
